// UniverCity-ui/src/app/Services/models/quiz-question.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface QuizQuestion {
  correctAnswer?: string;
  difficultyLevel?: number;
  option1?: string;
  option2?: string;
  option3?: string;
  option4?: string;
  question?: string;
  quizquestion_id?: number;
}