// UniverCity-ui/src/app/Services/models/quiz-sub-category.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { QuizCategory } from '../models/quiz-category';
import { QuizQuestion } from '../models/quiz-question';
export interface QuizSubCategory {
  name?: string;
  quizCategory?: QuizCategory;
  quizQuestions?: Array<QuizQuestion>;
  quizsubcategory_id?: number;
}